+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Procedure when bumping the version number:
 ### Changed
 - Respect colon-delimited emoji when calculating nick hue
 - Display colon-delimited emoji in nicks and messages
+- Non-export info is now printed to stderr instead of stdout
 
 ### Fixed
 - Mentions not being stopped by `>`