+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Procedure when bumping the version number:
 - Respect colon-delimited emoji when calculating nick hue
 - Display colon-delimited emoji in nicks and messages
 - Non-export info is now printed to stderr instead of stdout
+- Recognizes links without scheme (e. g. `euphoria.io` instead of `https://euphoria.io`)
 
 ### Fixed
 - Mentions not being stopped by `>`